third_party/blink/web_tests/wpt_internal/task-tracking/soft-navigation-heuristics/soft-navigation-detection.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<title>Detect simple soft navigation.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
</head>
<body>
  <main id=main>
  <a id=link>Click me!</a>
  </main>
  <script>
    promise_test(async t => {
      const link = document.getElementById("link");
      const clicked = new Promise(resolve => {
        link.addEventListener("click", async e => {
          // Jump through a task, to ensure task tracking is working properly.
          await new Promise(r => setTimeout(r, 0));

          // Fetch some content
          const response = await fetch("../resources/content.json");
          const json = await response.json();

          // Change the URL
          history.pushState({}, '', "/foobar.html");

          // Add the content to the main element
          const main = document.getElementById("main");
          const content = document.createTextNode(json["content"]);
          main.appendChild(content);

          resolve();
        });
      });
      if (test_driver) {
        test_driver.click(link);
      }
      await clicked;
      assert_equals(document.softNavigations, 1);
    }, "Test that a soft navigation is detected");
  </script>
</body>
</html>
